requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.3
python-multipart==0.0.9

# Semantic Chunking Dependencies
sentence-transformers>=2.2.0
nltk>=3.8.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
redis==5.0.4

# LLM & AI
langchain==0.2.5
langchain-community==0.2.5
langchain-core>=0.2.7,<0.3.0
langsmith==0.1.77
langgraph==0.1.5
langchain-openai==0.1.8

# Vector Store
chromadb==0.5.0
faiss-cpu==1.7.4
sentence-transformers==2.7.0  # For local embeddings

# Document Processing
python-docx==1.1.0  # For Word document support

# HTTP Client
httpx==0.27.0
aiohttp==3.9.5

# LLM Providers
openai==1.35.3  # For OpenAI provider compatibility

# Development
pytest==8.2.2
pytest-asyncio==0.23.7
black==24.4.2
ruff==0.4.9

# Evaluation & Metrics
ragas>=0.1.0
datasets>=2.14.0
evaluate>=0.4.0

# Visualization for Dashboard
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data Processing for Evaluation
pandas>=2.0.0
numpy>=1.24.0

# Logging & Monitoring
structlog==24.2.0

# Environment
python-dotenv==1.0.1